feat: update adaptive profile page
